// editor/src/renderer/components/source/source-editor-theme-types.ts
export type CodeEditorThemeId =
  | 'noveltea'
  | 'abcdef'
  | 'abyss'
  | 'androidStudio'
  | 'andromeda'
  | 'basicDark'
  | 'basicLight'
  | 'catppuccinMocha'
  | 'cobalt2'
  | 'forest'
  | 'githubDark'
  | 'githubLight'
  | 'gruvboxDark'
  | 'gruvboxLight'
  | 'highContrastDark'
  | 'highContrastLight'
  | 'materialDark'
  | 'materialLight'
  | 'materialOcean'
  | 'monokai'
  | 'nord'
  | 'palenight'
  | 'solarizedDark'
  | 'solarizedLight'
  | 'synthwave84'
  | 'tokyoNightDay'
  | 'tokyoNightStorm'
  | 'volcano'
  | 'vsCodeDark'
  | 'vsCodeLight';